feat: atomWithHash
