+++ src/utils.ts
@@ -11,4 +11,4 @@ export { freezeAtom, atomFrozenInDev } from './utils/freezeAtom'
 export { splitAtom } from './utils/splitAtom'
 export { atomWithDefault } from './utils/atomWithDefault'
 export { waitForAll } from './utils/waitForAll'
-export { urlHashAtom } from './utils/urlHashAtom'
+export { atomWithHash } from './utils/atomWithHash'
